requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
chromadb==0.4.22
google-genai==0.2.2
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
pymupdf==1.23.8
scikit-learn==1.4.0
networkx==3.2.1
requests==2.31.0
aiohttp==3.9.1
plotly==5.18.0